add group editable flag

+++ afrims/apps/groups/views.py
@@ -3,7 +3,7 @@ import logging
 from django import forms
 from django.http import HttpResponse
 from django.shortcuts import redirect, render_to_response, get_object_or_404
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, HttpResponseForbidden
 from django.core.paginator import Paginator, InvalidPage, EmptyPage
 from django.utils.translation import ugettext_lazy as _
 from django.template import RequestContext
@@ -38,6 +38,8 @@ def create_edit_group(request, group_id=None):
     group = None
     if group_id:
         group = get_object_or_404(Group, pk=group_id)
+        if not group.is_editable:
+            return HttpResponseForbidden('Access denied')
     if request.method == 'POST':
         form = GroupForm(request.POST, instance=group)
         if form.is_valid():
@@ -58,6 +60,8 @@ def create_edit_group(request, group_id=None):
 @transaction.commit_on_success
 def delete_group(request, group_id):
     group = get_object_or_404(Group, pk=group_id)
+    if not group.is_editable:
+        return HttpResponseForbidden('Access denied')
     if request.method == 'POST':
         group.delete()
         messages.info(request, 'Group successfully deleted')
@@ -127,4 +131,3 @@ def forwarding(request):
     }
     return render_to_response('groups/forwarding.html', context,
                               RequestContext(request))
-
